# server/src/middleware/__generated__/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"FirebaseFirestore.Timestamp": {
				"description": "A Timestamp represents a point in time independent of any time zone or\ncalendar, represented as seconds and fractions of seconds at nanosecond\nresolution in UTC Epoch time. It is encoded using the Proleptic Gregorian\nCalendar which extends the Gregorian calendar backwards to year one. It is\nencoded assuming all minutes are 60 seconds long, i.e. leap seconds are\n\"smeared\" so that no leap second table is needed for interpretation. Range\nis from 0001-01-01T00:00:00Z to 9999-12-31T23:59:59.999999999Z.",
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double",
						"description": "The number of seconds of UTC time since Unix epoch 1970-01-01T00:00:00Z."
					},
					"nanoseconds": {
						"type": "number",
						"format": "double",
						"description": "The non-negative fractions of a second at nanosecond resolution."
					}
				},
				"required": [
					"seconds",
					"nanoseconds"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"stripe.Stripe.Checkout.Session.Status": {
				"type": "string",
				"enum": [
					"complete",
					"expired",
					"open"
				]
			},
			"stripe.Stripe.Metadata": {
				"description": "Set of key-value pairs that you can attach to an object. This can be useful for storing additional information about the object in a structured format.",
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"type": "string"
				}
			},
			"MembershipTypeValues": {
				"enum": [
					"uoa_student",
					"non_uoa_student",
					"returning_member",
					"new_non_student"
				],
				"type": "string"
			},
			"MembershipPaymentResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stripeClientSecret": {
						"type": "string"
					},
					"membershipType": {
						"$ref": "#/components/schemas/MembershipTypeValues"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserMembershipRequestModel": {
				"properties": {
					"membershipType": {
						"$ref": "#/components/schemas/MembershipTypeValues"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"BookingPaymentResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stripeClientSecret": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UserBookingRequestingModel": {
				"properties": {
					"startDate": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					},
					"endDate": {
						"$ref": "#/components/schemas/FirebaseFirestore.Timestamp"
					}
				},
				"required": [
					"startDate",
					"endDate"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/payment/checkout_status": {
			"get": {
				"operationId": "GetCheckoutSessionDetails",
				"responses": {
					"200": {
						"description": "Session Fetched",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"metadata": {
											"$ref": "#/components/schemas/stripe.Stripe.Metadata"
										},
										"pricePaid": {
											"type": "number",
											"format": "double"
										},
										"customer_email": {
											"type": "string"
										},
										"status": {
											"$ref": "#/components/schemas/stripe.Stripe.Checkout.Session.Status"
										}
									},
									"required": [
										"metadata",
										"pricePaid",
										"customer_email",
										"status"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "sessionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/payment/membership": {
			"post": {
				"operationId": "GetMembershipPayment",
				"responses": {
					"200": {
						"description": "Session created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MembershipPaymentResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserMembershipRequestModel"
							}
						}
					}
				}
			}
		},
		"/payment/booking": {
			"post": {
				"operationId": "GetBookingPayment",
				"responses": {
					"200": {
						"description": "Created booking checkout session",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/BookingPaymentResponse"
								}
							}
						}
					}
				},
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserBookingRequestingModel"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}